Change IdentData data structures
In preparation for a rewrite of the mzIdentML parser, more documentation has
been added to the different IdentData data structures. In this commit there are
multiple breaking changes which affect parts of the code other than the
mzIdentML reader, including serialization, bindings, etc.

This commit won't compile. Additionally, the protein inference algorithm must be
reconsidered, and we need to prove it is working as expected.

The background for this series of changes is that the mzIdentML file is
organized differently depending on the source. This include missing fields and
differences in the implementation of the spec. For example, in 1.1.0 the
ProteinAmbiguityGroup does not exist, and thus we can't rely on it to generate a
list of proteins associated to a given PSM.