Add SpectrumMatch parsing to mzIdentML reader
Currenlty the mzIdenML reader is functional, but it is not currenly parsing the
PeptideModifications. This implementation works for both PeptideShaker->mzid and
Comet->ProteoWizard's idconvert->mzid. In principle it might work for other
cases, but we can't ensure that the retention_time is going to be present on
those.

For example, the mzIdentML file produced by Peaks or MS-GF+ does not contain the
retention time information. It will still be read, but we can't rely on that
information being there. This makes the linkage between ms2 events and PSM
impossible, as we can't also rely on the SpectrumIdentificationResult containing
the right `spectrumId` that matches the original MS2 id.