Improve performance of feature metamatch
General cleanup of this function has been performed. Switching from using
retention time to m/z for the binary search results in a big performance
increase (100x). The reason for this is that for any given retention time on a
ROI, we have many more distinct m/z points. The opposite is not true, and thus
with this change we minimize the amount of time that we spend on the linear
search part of the algorithm.